Configuration swagger and add endpoints

+++ src/main/java/com/aleksnose/hoteru/repository/IRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRepository extends CrudRepository<User, Integer> {
+public interface IRepository extends CrudRepository<User, Long> {
